2/22 21:39 java web prj update

+++ src/main/java/org/zerock/domain/MemberVO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class MemberVO {
 
-	private int mno;
+	private Long mno;
 	private String id;
 	private String password;
 	private String name;
